backend/requirements.txt
# Python dependencies for the GodWorld QR Store backend
fastapi==0.111.0
uvicorn[standard]==0.29.0
web3==6.18.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0          # async HTTP for Google Lens API calls
pillow==10.3.0         # image processing
pydantic==2.7.1
